naming
fixed folder stucture
messed with turtle bot code
