+++ .finalProject.wbproj
@@ -1,0 +1,12 @@
+Webots Project File version R2021b
+perspectives: 000000ff00000000fd0000000200000001000000f8000002bdfc0200000001fb0000001400540065007800740045006400690074006f00720100000000000002bd0000004500ffffff000000030000045a000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c01000000000000045a0000008700ffffff00000360000002bd00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff000000010000000200000119000004c90100000002010000000100
+sceneTreePerspectives: 000000ff000000010000000200000204000000fa0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: -1
+consoles: Console:All:All
+renderingDevicePerspectives: TIAGo Steel:camera;1;1;0;0
+renderingDevicePerspectives: TIAGo Steel:display;1;0.461111;0;0
+renderingDevicePerspectives: TIAGo Steel:range-finder;1;1;0;0
