--- finalProject.wbt
@@ -1,430 +1,0 @@
-#VRML_SIM R2021b utf8
-WorldInfo {
-  basicTimeStep 16
-  coordinateSystem "NUE"
-}
-Viewpoint {
-  orientation 0.532612698399652 -0.5920134746188139 -0.6048501958114919 4.112786402202066
-  position 4.679270015175837 18.76308061630109 9.899271790978315
-  follow "fence(2)"
-}
-TexturedBackground {
-}
-TexturedBackgroundLight {
-}
-CircleArena {
-  translation 0 0.01 0
-  radius 9
-  floorAppearance BrushedAluminium {
-  }
-  wallThickness 0.2
-  subdivision 40
-}
-Bench {
-  translation 0 0 -11.3
-}
-Bench {
-  translation 8.28524 0 -8.14847
-  rotation 0 -1 0 0.785401
-  name "bench(2)"
-}
-Bench {
-  translation -8.06275 0 -8.08448
-  rotation 0 -1 0 -0.7854053071795866
-  name "bench(1)"
-}
-Floor {
-  size 30 30
-  appearance Grass {
-  }
-}
-PaperBoat {
-  translation -3.15958 0 7.15964
-  rotation 0 1 0 -0.523595307179586
-  scale 2
-}
-PaperBoat {
-  translation -1.03535 0 7.50449
-  rotation 0 1 0 -0.261795307179586
-  name "paper boat(1)"
-  scale 2
-}
-StraightRoadSegment {
-  translation 21 0 -17
-  length 34
-}
-ControlledStreetLight {
-  translation 14.1 0 -14.09
-}
-ControlledStreetLight {
-  translation -14.22 0 -14.2
-  name "street light(1)"
-}
-SmallKiosk {
-  translation -6.04753 0 13.5895
-  rotation 0 1 0 -0.261795307179586
-}
-GlassRecyclingContainer {
-  translation -13.4806 0 -7.84753
-  rotation 0 1 0 -0.261795307179586
-}
-DrinkingFountain {
-  translation -7.91 0 12.11
-  rotation 0 1 0 -0.785395307179586
-}
-DrinkingFountain {
-  translation -11.3466 0 8.67332
-  rotation 0 1 0 -0.785395307179586
-  name "drinking fountain(1)"
-}
-SimpleTree {
-  translation 14 0 13
-  name "tree(7)"
-  height 6
-  radius 5
-  subdivision 7
-}
-Oak {
-  translation -8.31 0 -13.78
-  name "oak tree(4)"
-}
-Oak {
-  translation -9.15 0 -11.77
-  name "oak tree(7)"
-}
-Oak {
-  translation -10.87 0 -1.44
-  rotation 0 1 0 0.785398
-  name "oak tree(5)"
-}
-Oak {
-  translation 13.6885 0 -0.952373
-  rotation 0 1 0 0.785398
-  name "oak tree(6)"
-}
-Oak {
-  translation 2.52 0 13.32
-  rotation 0 1 0 1.5708
-  name "oak tree(3)"
-}
-Ball {
-  translation 8.7 0.1 10.79
-  radius 0.1
-}
-Slide {
-  translation 6.56863 0 9.54728
-  rotation 0 -1 0 0.785397
-}
-RectangleArena {
-  translation 9.53572 0.01 9.33578
-  rotation 0 1 0 -0.785395307179586
-  floorSize 6 7
-  floorAppearance Sand {
-  }
-}
-TurtleBot3Burger {
-  translation 3.92 0 11.22
-  controller "finalTurtlebot"
-}
-Oak {
-  translation 2.52 0 -11.95
-}
-Oak {
-  translation 11.25 0 -12.95
-  name "oak tree(8)"
-}
-Oak {
-  translation 13.48 0 -9.99
-  name "oak tree(9)"
-}
-Oak {
-  translation 12.29 0 6.07
-  rotation 0 1 0 0.785398
-  name "oak tree(2)"
-}
-Oak {
-  translation -12.23 0 11.76
-  name "oak tree(1)"
-}
-Oak {
-  translation -13.58 0 9.21
-  name "oak tree(10)"
-}
-Rock17cm {
-  translation 9.29 0 0
-}
-Rock17cm {
-  translation 7.65 0 -5.33
-  name "rock 17 cm(2)"
-}
-Rock17cm {
-  translation 9.25 0 0.56
-  rotation 0 1 0 1.5708
-  name "rock 17 cm(3)"
-}
-Rock17cm {
-  translation 7.66 0 5.31
-  rotation 0 1 0 -2.094395307179586
-  name "rock 17 cm(4)"
-}
-Rock17cm {
-  translation 8.89576 0 -2.81362
-  rotation 0 1 0 1.8326
-  name "rock 17 cm(1)"
-}
-StoneFountain {
-}
-Floor {
-  translation 16 0 0
-  rotation 0 -1 0 0
-  name "floor(1)"
-  size 2 34
-  appearance Pavement {
-  }
-}
-Floor {
-  translation -16 0 0
-  rotation 0 -1 0 0
-  name "floor(2)"
-  size 2 34
-  appearance Pavement {
-  }
-}
-Floor {
-  translation 0 0 16
-  rotation 0 -1 0 1.5707996938995747
-  name "floor(3)"
-  size 2 30
-  appearance Pavement {
-  }
-}
-Floor {
-  translation 0 0 -16
-  rotation 0 -1 0 1.5707996938995747
-  name "floor(4)"
-  size 2 30
-  appearance Pavement {
-  }
-}
-Swing {
-  translation 11.232 0 9.45703
-  rotation 0 -1 0 1.8326
-}
-Fence {
-  translation -15 0 14.92
-  poleGap 0.3
-  path [
-    0 0 0
-    30 0 0
-  ]
-}
-Fence {
-  translation -15 0 -14.9
-  name "fence(1)"
-  poleGap 0.3
-  path [
-    0 0 0
-    30 0 0
-  ]
-}
-Fence {
-  translation -15 0 -14.9
-  rotation 0 1 0 -1.5707996938995747
-  name "fence(2)"
-  poleGap 0.3
-  path [
-    0 0 0
-    30 0 0
-  ]
-}
-Fence {
-  translation 14.96 0 -14.9
-  rotation 0 1 0 -1.5707996938995747
-  name "fence(3)"
-  poleGap 0.3
-  path [
-    0 0 0
-    30 0 0
-  ]
-}
-PublicBin {
-  translation 9.44 0 -6.77
-}
-PublicBin {
-  translation -9.42 0 -6.55
-  name "public bin(3)"
-}
-PublicBin {
-  translation 11.77 0 5.09
-  name "public bin(1)"
-}
-PublicBin {
-  translation -4.2 0 13.33
-  name "public bin(2)"
-}
-SimpleTree {
-  translation -12 0 -11
-  height 6
-  radius 4
-  subdivision 6
-}
-SimpleTree {
-  translation -10 0 14
-  name "tree(6)"
-  height 5
-  radius 4
-  subdivision 6
-}
-SimpleTree {
-  translation 13 0 0
-  name "tree(8)"
-  type "white pine"
-  height 1
-  radius 1
-  subdivision 5
-}
-SimpleTree {
-  translation -13 0 0
-  name "tree(7)"
-  type "crab apple tree"
-  height 5
-  radius 3
-  subdivision 5
-}
-SimpleTree {
-  translation -10 0 11
-  name "tree(7)"
-  type "crab apple tree"
-  height 5
-  radius 3
-  subdivision 5
-}
-SimpleTree {
-  translation 8 0 -12
-  name "tree(7)"
-  type "crab apple tree"
-  height 5
-  radius 3
-  subdivision 5
-}
-SimpleTree {
-  translation 14 0 3
-  name "tree(7)"
-  type "crab apple tree"
-  height 5
-  radius 3
-  subdivision 5
-}
-SimpleTree {
-  translation 3 0 -11
-  name "tree(7)"
-  type "white pine"
-  height 1
-  radius 1
-  subdivision 5
-}
-SimpleTree {
-  translation -3 0 -11
-  name "tree(7)"
-  type "white pine"
-  height 1
-  radius 1
-  subdivision 5
-}
-SimpleTree {
-  translation -6 0 -10
-  name "tree(7)"
-  type "white pine"
-  height 1
-  radius 1
-  subdivision 5
-}
-SimpleTree {
-  translation 12 0 -12
-  name "tree(7)"
-  type "white pine"
-  height 1
-  radius 1
-  subdivision 5
-}
-SimpleTree {
-  translation 13 0 -14
-  name "tree(7)"
-  type "white pine"
-  height 2
-  subdivision 5
-}
-SimpleTree {
-  translation 14 0 -11
-  name "tree(7)"
-  type "white pine"
-  height 2
-  subdivision 5
-}
-SimpleTree {
-  translation -14 0 -14
-  name "tree(7)"
-  type "white pine"
-  height 2
-  subdivision 5
-}
-SimpleTree {
-  translation -14 0 -10
-  name "tree(7)"
-  type "white pine"
-  height 2
-  subdivision 5
-}
-SimpleTree {
-  translation -11 0 -13
-  name "tree(7)"
-  type "white pine"
-  height 2
-  subdivision 5
-}
-SimpleTree {
-  translation 7 0 14
-  name "tree(7)"
-  type "white pine"
-  height 2
-  subdivision 5
-}
-SimpleTree {
-  translation 12 0 13
-  name "tree(7)"
-  type "white pine"
-  height 2
-  subdivision 5
-}
-TiagoSteel {
-  translation 2.58032 0.111331 10.0546
-  rotation 0.9310809087605929 0.2756919729841178 0.23891897658761382 -1.6198753071795862
-  controller "final"
-  cameraSlot [
-    Display {
-      width 900
-      height 900
-    }
-    Compass {
-      translation -0.206 0 -0.041438
-    }
-    GPS {
-      translation -0.190515 0 -0.015552
-    }
-    Camera {
-      rotation 0 1 0 -1.57
-      width 320
-      height 240
-      recognition Recognition {
-        segmentation TRUE
-      }
-    }
-    RangeFinder {
-      rotation 0 1 0 -1.57
-      width 320
-      height 240
-      minRange 0.1
-      maxRange 5.6
-    }
-  ]
-}
