+++ protos/.Lab4RobotisLds01.cache
@@ -1,0 +1,11 @@
+Webots Proto Cache File version 1.5
+protoFileHash: 5ac819ec7e751a395d6992c0fb190333
+needsRobotAncestor: 1
+baseType: Lidar
+slotType: 
+parameters: translation,rotation,name,noise,enablePhysics
+tags: static
+license:  Copyright Cyberbotics Ltd. Licensed for use only with Webots.
+licenseUrl:  https://cyberbotics.com/webots_assets_license
+documentationUrl: https://www.cyberbotics.com/doc/guide/lidar-sensors#lds-01
+info: Model\sof\sthe\sRobotis\sLDS-01\sLaser\sDistance\sSensor\nReference:\shttp://www.robotis.us/360-laser-distance-sensor-lds-01-lidar/\n
