+++ worlds/.finalProject.wbproj
@@ -1,0 +1,13 @@
+Webots Project File version R2021b
+perspectives: 000000ff00000000fd0000000200000001000003a6000002bbfc0200000001fb0000001400540065007800740045006400690074006f00720100000000000002bb0000008b00ffffff0000000300000780000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000007800000008700ffffff000003d8000002bb00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff000000010000000200000118000002be0100000002010000000100
+sceneTreePerspectives: 000000ff00000001000000020000027e000000000100000002010000000200
+minimizedPerspectives: 000000ff00000000fd0000000200000001000003a6000002bbfc0200000001fb0000001400540065007800740045006400690074006f00720100000000000002bb0000008b00ffffff0000000300000780000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000007800000008700ffffff000003d8000002bb00000001000000020000000100000008fc00000000
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: 0 "controllers/finalTurtleBot/finalTurtlebot.py" "controllers/tiago/tiago.py"
+consoles: Console:All:All
+renderingDevicePerspectives: TIAGo Steel:camera;0;1;0;0
+renderingDevicePerspectives: TIAGo Steel:display;0;0.315556;0;0
+renderingDevicePerspectives: TIAGo Steel:range-finder;0;1;0;0
