+++ worlds/TiagoSteel.wbo
@@ -1,0 +1,33 @@
+#VRML_OBJ R2021b utf8
+TiagoSteel {
+  translation 4.46793 0.172269 8.05674
+  rotation 0.9310809087605929 0.2756919729841178 0.23891897658761382 -1.6198753071795862
+  controller "lab5_controller"
+  cameraSlot [
+    Display {
+      width 360
+      height 360
+    }
+    Compass {
+      translation -0.206 0 -0.041438
+    }
+    GPS {
+      translation -0.190515 0 -0.015552
+    }
+    Camera {
+      rotation 0 1 0 -1.57
+      width 320
+      height 240
+      recognition Recognition {
+        segmentation TRUE
+      }
+    }
+    RangeFinder {
+      rotation 0 1 0 -1.57
+      width 320
+      height 240
+      minRange 0.1
+      maxRange 5.6
+    }
+  ]
+}
